Split HRParser in two parts to enhance extensibility.
- pratt.py  : the scaffolding needed for the parser
- parser.py : the actual parser for HR
